(PC-21051)[PRO] refactor: protection of private routes
