--- pro/src/app/App/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useIsRoutePublic } from './useIsRoutePublic'
