--- pro/src/app/App/hooks/useIsRoutePublic/index.ts
@@ -1,1 +1,0 @@
-export { default as useIsRoutePublic } from './useIsRoutePublic'
